replace autocomplete with autosuggest

+++ web/src/components/ShoppingCart/AddCartItem.tsx
@@ -1,6 +1,6 @@
-import React, { useState, useContext } from 'react'
+import React, { useState } from 'react'
 import Form from 'react-bootstrap/Form'
-import Autocomplete from 'react-autocomplete'
+import Autosuggest from 'react-autosuggest';
 import Col from 'react-bootstrap/Col'
 import Button from 'react-bootstrap/Button'
 import { FaPlus } from 'react-icons/fa'
@@ -8,6 +8,7 @@ import { CartItemInterface, User } from 'lib/interfaces'
 import useForm from 'lib/useForm'
 import { ADD_CART_ITEM, GET_CART_ITEMS_BY_USER_ID, DELETE_CART_ITEM } from 'graphql/queriesAndMutations'
 import { useMutation } from '@apollo/react-hooks'
+import moduleStyles from './AddCartItem.module.css'
 
 interface Props {
     itemSuggestions: CartItemInterface[]
@@ -28,12 +29,10 @@ const formControlStyles = {
     backgroundClip: 'padding-box',
     border: '1px solid #ced4da',
     borderRadius: '0.4rem',
-    transition: 'border-color 0.15s'
+    transition: 'border-color 0.15s',
 }
 
 const AddCartItem: React.FC<Props> = ({ itemSuggestions, me }) => {
-    const [selectedItem, setSelectedItem] = useState<CartItemInterface | null>(null)
-
     const [addCartItem] = useMutation(ADD_CART_ITEM, {
         refetchQueries: [{ query: GET_CART_ITEMS_BY_USER_ID, variables: { id: me.id } }]
     })
@@ -59,17 +58,61 @@ const AddCartItem: React.FC<Props> = ({ itemSuggestions, me }) => {
         resetForm()
     }
 
-    const handleSelection = (val: any) => {
+
+    const handleDelete = (id: number|undefined) => {
+        if (id && window.confirm('Are you sure you want to delete item?'))
+            deleteCartItem({ variables: { id } })
+    }
+
+    // AUTOSUGGEST
+
+    const [suggestions, setSuggestions] = useState<CartItemInterface[]>(itemSuggestions)
+
+    // Teach Autosuggest how to calculate suggestions for any given input value.
+    const getSuggestions = (value: string) => {
+        const inputValue = value.trim().toLowerCase();
+        const inputLength = inputValue.length;
+
+        return inputLength === 0 ? [] : itemSuggestions.filter(item =>
+            item.name.toLowerCase().slice(0, inputLength) === inputValue
+        );
+    };
+
+    // When suggestion is clicked, Autosuggest needs to populate the input
+    // based on the clicked suggestion.
+    const getSuggestionValue = (suggestion: CartItemInterface) => (suggestion.name)
+
+    const renderSuggestion = (item: CartItemInterface, params: {isHighlighted: boolean, query: string}) => (        
+        <div style={{ background: params.isHighlighted ? 'lightgray' : 'white'}}>
+            {item.name} | {item.amount} | {item.unit} | {item.aisle} | <Button onClick={() => handleDelete(item.id)} size="sm" variant="danger">x</Button>
+        </div>
+    );
+
+    // Autosuggest will call this function every time you need to update suggestions.
+    const onSuggestionsFetchRequested = ({ value }: { value: any }) => {
+        setSuggestions(getSuggestions(value))
+        console.log(value);
+
+    };
+
+    const onSuggestionsClearRequested = () => {
+        setSuggestions([])
+    };
+
+    const onSuggestionSelected = (event: any, { suggestion }: {suggestion: CartItemInterface}) => {
         // set other input values once an item name is selected 
-        const { name, amount, unit, aisle } = selectedItem!
+        const { name, amount, unit, aisle } = suggestion!
         const update: any = { name, amount, unit, aisle }
         forceChange(update)
     }
 
-    const handleDelete = (id: number) => {
-        if (window.confirm('Are you sure you want to delete item?'))
-            deleteCartItem({ variables: { id } })
-    }
+    const inputProps = {
+        placeholder: 'Enter item name',
+        value: inputs.name,
+        onChange: handleChange,
+        style: formControlStyles, 
+        name: 'name'
+    };
 
     return (
         <Form onSubmit={handleSubmit}>
@@ -77,31 +120,28 @@ const AddCartItem: React.FC<Props> = ({ itemSuggestions, me }) => {
                 <Form.Row>
                     <Col>
                         <Form.Label>Item Name</Form.Label>
-                        <Autocomplete
-                            getItemValue={(item) => { setSelectedItem(item); return item.name }}
-                            items={itemSuggestions} renderItem={(item, isHighlighted) =>
-                                <div style={{ background: isHighlighted ? 'lightgray' : 'white' }}>
-                                    {item.name} | {item.amount} | {item.unit} | {item.aisle} | <Button onClick={() => handleDelete(item.id)} size="sm" variant="danger">x</Button>
-                                </div>
-                            }
-                            inputProps={{ style: formControlStyles, name: 'name' }}
-                            // selectOnBlur={true}
-                            value={inputs.name}
-                            onChange={handleChange}
-                            onSelect={handleSelection}
+                        <Autosuggest
+                            suggestions={suggestions}
+                            onSuggestionsFetchRequested={onSuggestionsFetchRequested}
+                            onSuggestionsClearRequested={onSuggestionsClearRequested}
+                            getSuggestionValue={getSuggestionValue}
+                            onSuggestionSelected={onSuggestionSelected}
+                            renderSuggestion={renderSuggestion}
+                            inputProps={inputProps}
                         />
+
                     </Col>
                     <Col>
                         <Form.Label>Amount</Form.Label>
-                        <Form.Control type="number" name="amount" value={inputs.amount} onChange={handleChange} min="1" />
+                        <Form.Control className={moduleStyles.formControl} type="number" name="amount" value={inputs.amount} onChange={handleChange} min="1" />
                     </Col>
                     <Col>
                         <Form.Label>Units</Form.Label>
-                        <Form.Control type="text" name="unit" value={inputs.unit} onChange={handleChange} placeholder="lbs/bags/etc" />
+                        <Form.Control className={moduleStyles.formControl} type="text" name="unit" value={inputs.unit} onChange={handleChange} placeholder="lbs/bags/etc" />
                     </Col>
                     <Col>
                         <Form.Label>Aisle #</Form.Label>
-                        <Form.Control type="number" name="aisle" value={inputs.aisle} onChange={handleChange} min="0" />
+                        <Form.Control className={moduleStyles.formControl} type="number" name="aisle" value={inputs.aisle} onChange={handleChange} min="0" />
                     </Col>
                     <Col>
                         <Button type="submit"><FaPlus /></Button>
